requirements.txt
faiss-cpu
numpy
requests
tqdm
fastapi